backend/requirements.txt
fastapi[standard]>=0.68.0
python-dotenv>=0.19.0
psycopg2-binary>=2.9.1